Move the preparation the node environment from NodeRuntime to NodeEnvironment
